+++ node/app/controllers/link.server.controller.js
@@ -4,6 +4,7 @@ require('../models/statistics.server.model');
 var Link = mongoose.model("Link");
 var url = require('url');
 var statistic = require('../../app/controllers/statistic.server.controller');
+var linkservice = require('../services/link.server.service');
 
 var createLinkModel = function (link) {
     var LinkSuccess = {
@@ -17,38 +18,21 @@ var createLinkModel = function (link) {
     LinkSuccess.shortURL = GLOBAL_SERVER + '/' + link.shortlink;
     LinkSuccess.shortQR = GLOBAL_SERVER + '/qr/' + link.shortlink;
     LinkSuccess.longLink = url.parse(link.longlink).hostname;
-    LinkSuccess.longURL = link.longlink;    
+    LinkSuccess.longURL = link.longlink;
     return LinkSuccess;
 };
 
 // fixes "createLinkModel is not defined" error
-exports.createLinkModel = createLinkModel; 
+exports.createLinkModel = createLinkModel;
 
 exports.create = function (req, res, next) {
-    var link = new Link(req.body);    
-    var LinkSuccess = createLinkModel(link);  
-    link.save(function (err) {
-        if (err) {
-            return next(err);
-        } else {
-            res.json(LinkSuccess);
-            res.end();
-            //init statistic
-            statistic.initStatistics(link.shortlink);
-            statistic.initQRStatistics(link.shortlink);
-
-        }
-    });
+    var link = new Link(req.body);
+    var LinkSuccess = createLinkModel(link);
+    linkservice.save(req, res, next, link, LinkSuccess);
 };
 
 exports.list = function (req, res, next) {
-    Link.find({}, function (err, links) {
-        if (err) {
-            return next(err);
-        } else {
-            res.json(links);
-        }
-    });
+    linkservice.list(req, res, next);
 };
 
 
@@ -71,36 +55,11 @@ exports.redirectQR = function (req, res) {
 };
 
 exports.linkByShort = function (req, res, next, slink) {
-    Link.findOne({
-            'shortlink': slink
-        },
-        function (err, link) {
-            if (err) {
-                console.log(err);
-                return next(err);
-            } else {
-                if (link == null) {
-                    res.status(404).render('404', {
-                        title: "Sorry, page not found"
-                    });
-                    res.end();
-                } else {
-                    req.body = link;
-                    next();
-                }
-            }
-        }
-    );
+    linkservice.linkByShort(req, res, next, slink);
 };
 
 exports.update = function (req, res, next) {
-    Link.findByIdAndUpdate(req.link.id, req.body, function (err, link) {
-        if (err) {
-            return next(err);
-        } else {
-            res.json(link);
-        }
-    });
+      linkservice.update(req, res, next);
 };
 
 exports.validateURL = function (req, res, next) {
@@ -143,26 +102,7 @@ exports.findLongLink = function (req, res, next) {
         longURL: String
     };
     if (GLOBAL_PREMIUM == false || link.shortlink == '') {
-        Link.find({
-                "longlink": link.longlink
-            },
-            function (err, docs) {
-                if (docs.length != 0) {
-                    link.shortlink = docs[0].shortlink;
-
-                    LinkSuccess.shortLink = '/' + link.shortlink;
-                    LinkSuccess.shortURL = GLOBAL_SERVER + '/' + link.shortlink;
-                    LinkSuccess.shortQR = GLOBAL_SERVER + '/qr/' + link.shortlink;
-                    LinkSuccess.longLink = url.parse(link.longlink).hostname;
-                    LinkSuccess.longURL = link.longlink;
-
-                    res.json(LinkSuccess);
-                    res.end();
-                } else {
-                    next();
-                }
-            }
-        );
+        linkservice.findLongLink(req, res, next, link, LinkSuccess);
     } else {
         next();
     }
@@ -217,47 +157,7 @@ exports.checkShortLink = function (req, res, next) {
         link.shortlink = randomText();
     }
 
-    Link.find({
-            "shortlink": link.shortlink
-        },
-        function (err, docs) {
-            if (docs.length != 0) {
-                //ShortLink does exist, search new one or throw error
-                found = true;
-                if (GLOBAL_PREMIUM == 'true') {
-                    LinkError.shortlink = link.shortlink;
-                    LinkError.longlink = link.longlink;
-                    LinkError.error = "Shortlink " + link.shortlink + " is already taken";
-
-                    res.json(LinkError);
-                    res.end();
-                } else {
-                    while (found == true) {
-                        Link.find({
-                                "shortlink": link.shortlink
-                            },
-                            function (err, docs) {
-                                if (docs.length != 0) {
-                                    link.shortlink = randomText();
-                                    found = true;
-                                } else {
-                                    found = false;
-                                }
-
-                            });
-                    }
-
-                    req.body = link;
-                    next();
-                }
-
-            } else {
-
-                req.body = link;
-                next();
-            }
-        }
-    );
+   linkservice.checkShortLink(req, res, next, link, LinkError);
 };
 
 var randomText = function() {
