+++ node/test/index.server.controller.js
@@ -5,6 +5,7 @@ var index = require("../app/controllers/index.server.controller.js");
 describe("Index Controller", function () {
     describe("Check Title", function () {
         it("checks the title of the page", function () {
+          GLOBAL_PREMIUM=false;
             var req = httpMocks.createRequest();
             var res = httpMocks.createResponse();
             index.render(req, res);
